Extract Agent conversation error detail formatter
